refactor: move CLI/tool logic into internal packages for testability
- Move slog-smoke logic to internal/slogsmoke with testable functions
- Move dagcheck visualization logic to internal/dagcheck with injectable ExecCommand
- Move verify_matrix logic to internal/verifymatrix with configurable runner
- Keep main functions thin - they only handle flags and call internal packages
- All binaries maintain same CLI interface and behavior
- Enable unit testing of previously untestable logic
