+++ scripts/coverage.out
@@ -1,0 +1,82 @@
+mode: set
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:92.13,100.87 5 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:100.87,101.17 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:101.17,103.4 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:106.3,106.91 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:106.91,108.4 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:111.3,111.77 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:111.77,113.4 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:115.3,118.33 3 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:118.33,119.29 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:119.29,121.5 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:121.10,121.38 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:121.38,123.5 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:126.3,126.13 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:129.2,129.16 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:129.16,131.3 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:134.2,145.20 8 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:145.20,148.3 2 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:148.8,148.44 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:148.44,151.3 2 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:151.8,154.3 2 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:157.55,161.16 3 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:161.16,164.3 2 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:165.2,170.21 4 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:170.21,175.52 3 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:175.52,176.44 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:176.44,178.25 2 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:178.25,181.23 3 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:181.23,183.7 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:189.3,189.44 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:189.44,191.54 2 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:191.54,192.21 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:192.21,201.6 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:206.3,206.89 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:206.89,215.4 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:219.2,219.16 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:222.120,228.24 3 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:228.24,229.41 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:229.41,231.4 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:231.9,231.46 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:231.46,233.4 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:237.2,237.50 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:237.50,246.3 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:248.2,248.50 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:248.50,257.3 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:259.2,259.55 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:259.55,268.3 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:270.2,270.12 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:273.41,280.66 3 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:280.66,284.43 3 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:284.43,288.34 3 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:288.34,289.22 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:289.22,291.6 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:291.11,291.29 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:291.29,293.6 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:294.5,294.13 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:297.4,297.31 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:297.31,299.5 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:299.10,299.38 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:299.38,301.5 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:302.4,302.12 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:307.2,307.13 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:307.13,309.3 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:309.8,309.20 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:309.20,311.3 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:312.2,312.10 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:315.47,316.23 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:316.23,319.3 2 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:322.2,326.28 4 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:326.28,327.21 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:328.16,329.30 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:330.18,331.34 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:332.15,333.26 1 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:338.2,338.21 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:338.21,340.28 2 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:340.28,342.4 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:346.2,346.23 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:346.23,348.30 2 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:348.30,350.4 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:354.2,354.35 1 1
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:354.35,356.26 2 0
+github.com/nephio-oran-claude-agents/scripts/verify_matrix.go:356.26,358.4 1 0
